First simple version for plotting modes files.
Now the Mode object has a function called plot_schedule which goal is to
plot how the modes evolves as the time goes by using the timelines in a
modes file.

For now we can only see when there is activity in the plot but it is all in
the same colour. The purpose of this plot is to plot with different colors
depending on the mode it is using.