4th compatibility patch.
Epys can now handle and plot module_states files. I have also fixed a
test_data file that had some errors.

This module_states files are compatible with the last compatibility patch
functions and it works using the same classes and methods. Perhaps in the
future we will have to differenciate them. But for now this is the best
approach.