Performance improvement.
Now modes_schedule plot function takes only 2.5% of the time it took at the
beginning. It goes on average 40 times faster. It can read and plot a 35000
lines file in 10-15 sec on my laptop. We think it is enough for now, if we
need a better performance in the future we will think how to solve this.