First prototype for the ITL object.
This first prototype should contain the same function that the last EVF file
model contains.

I also have some doubts of the meaning of some paramenters in the ITL file,
how they are used and how to interpret them for a future work.

Although the code is commmented, documentation for autodoc is missing in this
class for now.

To_file function is not adapted yet, because I want to discuss with Jonathan
about the use of this function.