Modularity and modes recognition.
The code is more modular now and more code is reused. Depending on how we are
going to handle some of the variables in the different tables we could even
use the same code for reading every table, for now I will copy and paste the
same code because maybe in the future we will need to add special code
depending on every table's characteristics.

Now the parser also loads the mode tables.